+++ cl/corpus_importer/utils.py
@@ -1,5 +1,5 @@
 from datetime import date
-from typing import Any, Optional
+from typing import Any, Iterator, List, Optional
 
 import bs4
 from django.utils.timezone import now
@@ -46,6 +46,117 @@ def get_start_of_quarter(d: Optional[date] = None) -> date:
     return max([q for q in quarter_dates if q <= d])
 
 
+def make_subset_range(cl_characters: str, max_string: str) -> List[int]:
+    """Find indices for matching max_string in CL opinion
+
+    :param cl_characters: The stripped down CL characters
+    :param max_string: The current largest identified substring
+    :return: Range of indices of match to CL as list
+    """
+    string_index_start = cl_characters.find(max_string)
+    string_index_end = string_index_start + len(max_string)
+    return list(range(string_index_start, string_index_end))
+
+
+def filter_subsets(lists: List[List[int]]) -> Iterator[List[int]]:
+    """Filter subsets from matches
+
+    Given list of lists, return new list of lists without subsets
+
+    :param lists: List of matched lists ranges
+    :return: Reduced list of matches
+    """
+
+    for match in lists:
+        if not any(
+            is_subset(match, other_matches)
+            for other_matches in lists
+            if match is not other_matches
+        ):
+            yield match
+
+
+def is_subset(match: List[int], other_match: List[int]) -> bool:
+    """Check if match is a subset of other matches
+
+    Check if needle is ordered subset of haystack in O(n)
+    :param match: Matching range of text as the indices
+    :param other_match: Other matching range of text as indices
+    :return: Is match a subset of other match
+    """
+
+    if len(other_match) < len(match):
+        return False
+    index = 0
+    for element in match:
+        try:
+            index = other_match.index(element, index) + 1
+        except ValueError:
+            return False
+    else:
+        return True
+
+
+def compare_documents(harvard_characters: str, cl_characters: str) -> int:
+    """Compare Harvard text to CL opinion text
+
+    This code iterates over two opinions logging similar stretches and then
+    returns a percentage of the total overlapping characters
+
+    :param harvard_characters: The stripped down opinion text from Harvard
+    :param cl_characters: The stripped down opinion text on Courtlistener
+    :return: Percentage (as integer) overlapping content
+    """
+
+    start, stop, count = 0, 0, 0
+    matched_substring = ""
+    found_overlaps = []
+    while stop < len(harvard_characters):
+        stop += 1
+        harvard_substring = harvard_characters[start:stop]
+        if harvard_substring in cl_characters:
+            matched_substring = harvard_substring
+        else:
+            if len(matched_substring) > 5:
+                subset = make_subset_range(cl_characters, matched_substring)
+                found_overlaps.append(subset)
+            matched_substring = ""
+            start = stop - 1
+    if len(matched_substring) > 5:
+        subset = make_subset_range(cl_characters, matched_substring)
+        found_overlaps.append(subset)
+
+    # If we checked our subsets as we parsed-we wouldn't need to do this
+    # filtering here. This is a good candidate for refactoring.
+    filtered_subset = list(filter_subsets(found_overlaps))
+    for overlap in filtered_subset:
+        count += len(overlap)
+
+    percent_match = int(
+        100 * (count / min([len(harvard_characters), len(cl_characters)]))
+    )
+    return percent_match
+
+
+def wrap_text(length, text):
+    """Wrap text to specified length without cutting words
+    :param length: max length to wrap
+    :param text: text to wrap
+    :return: text wrapped
+    """
+    words = text.split(" ")
+    if words:
+        lines = [words[0]]
+        for word in words[1:]:
+            if len(lines[-1]) + len(word) < length:
+                lines[-1] += f" {word}"
+            else:
+                lines.append(word)
+                break
+        return " ".join(lines)
+    return ""
+
+
 def similarity_scores(
     texts_to_compare_1: list[str], texts_to_compare_2: list[str]
 ) -> list[list[float]]:
@@ -87,9 +198,9 @@ def match_lists(
     :return: Matches if found or False
     """
     # We import this here to avoid a circular import
-    from cl.corpus_importer.management.commands.harvard_opinions import (
-        compare_documents,
-    )
+    # from cl.corpus_importer.management.commands.harvard_opinions import (
+    #     compare_documents,
+    # )
 
     # Convert harvard HTML to Text to compare
     harvard_opinions_list = [h.getText() for h in harvard_opinions_list]
@@ -119,22 +230,3 @@ def match_lists(
         matches[i] = j
 
     return matches
-
-
-def wrap_text(length, text):
-    """Wrap text to specified length without cutting words
-    :param length: max length to wrap
-    :param text: text to wrap
-    :return: text wrapped
-    """
-    words = text.split(" ")
-    if words:
-        lines = [words[0]]
-        for word in words[1:]:
-            if len(lines[-1]) + len(word) < length:
-                lines[-1] += f" {word}"
-            else:
-                lines.append(word)
-                break
-        return " ".join(lines)
-    return ""
